% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{collect_3d}
\alias{collect_3d}
\title{Translate 2D mosaic information to 3D attributes}
\usage{
collect_3d(image_list, mosaic_height = 6, highest_el = "light")
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_bricks]{collect_bricks()}}. Note that
only flat mosaics are supported.}

\item{mosaic_height}{Height of bricks in image. Default is 6.}

\item{highest_el}{Color of highest elevation bricks. Supported values are
\code{light} (default) and \code{dark}.}
}
\value{
list with \code{image_list} contents augmented with following components:
\itemize{
\item \code{threed_elevation}
\item \code{threed_hillshade}
}
}
\description{
Translate 2D mosaic information to 3D attributes
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks() \%>\%
  collect_3d()

}
