% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{collect_bricks}
\alias{collect_bricks}
\title{Title}
\usage{
collect_bricks(image_list, mosaic_type = "flat")
}
\arguments{
\item{image_list}{list object containing the \code{Img_scaled} data frame
as produced by the \code{\link[=legoize]{legoize()}} function.}

\item{mosaic_type}{String for mosaic type to produce. A \code{flat} mosaic
places a single layer of LEGO plates on a base plate with study-side up.
A \code{stacked} mosaic staggers bricks and places them horizontally.}
}
\value{
list with following components:
\itemize{
\item \code{Img_bricks}: Updated LEGO image
\item \code{ID_bricks}: Data frame with brick IDs
\item \code{mosaic_type}: Type of mosaic specfied by user
\item \code{pieces}: Data frame with number of bricks needed`
}
}
\description{
Title
}
\examples{
img <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
res <- scale_image(img, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
}
