% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{col_12}
\alias{col_12}
\alias{col_6}
\alias{col_4}
\alias{col_3}
\alias{col_2}
\title{Columns 12, 6 and 4}
\usage{
col_12(...)

col_6(...)

col_4(...)

col_3(...)

col_2(...)
}
\description{
Most shiny columns are 12, 6 or 4 of width.
These are convenient wrappers around
\code{column(12, ...)}, \code{column(6, ...)} and \code{column(4, ...)}.
}
