% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{count_bricks}
\alias{count_bricks}
\title{Utility function to create tidy pieces data frame tailored to specific steps}
\usage{
count_bricks(df, step_id, mosaic_type = "flat")
}
\arguments{
\item{df}{data frame produced by \code{\link[=generate_steps]{generate_steps()}}}

\item{step_id}{which step to filter. Format is \code{Step 01}, etc.}

\item{mosaic_type}{String for mosaic type to produce. A \code{flat} mosaic
places a single layer of LEGO plates on a base plate with study-side up.
A \code{stacked} mosaic staggers bricks and places them horizontally.}
}
\value{
data frame corresponding to same format as the \code{pieces} data frame
produced by \code{\link[=collect_bricks]{collect_bricks()}}
}
\description{
Utility function to create tidy pieces data frame tailored to specific steps
}
