% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{undisplay}
\alias{undisplay}
\alias{display}
\title{Hide or display a tag}
\usage{
undisplay(tag)

display(tag)
}
\arguments{
\item{tag}{the tag}
}
\value{
a tag
}
\description{
Hide or display a tag
}
\examples{
\dontrun{
## Hide
a <- shiny::tags$p(src = "plop", "pouet")
undisplay(a)
b <- shiny::actionButton("go_filter", "go")
undisplay(b)
}
\dontrun{
## Show
a <- shiny::tags$p(src = "plop", "pouet")
a_hidden <- undisplay(a)
display(a_hidden)
# do not change not hidden tags
b_show <- shiny::actionButton("go_filter", "go")
display(b_show)
# Keep other attributes
b_show$attribs$style <- 'display: none; background: red'
display(b_show)
}
}
