% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{display_3d}
\alias{display_3d}
\title{Display 3D render of mosaic}
\usage{
display_3d(image_list, ...)
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_3d]{collect_3d()}}}

\item{...}{Additional parameters passed to \code{\link[rayshader:plot_3d]{rayshader::plot_3d()}}}
}
\description{
Display 3D render of mosaic
}
