% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{display_pieces}
\alias{display_pieces}
\title{Visualize required pieces for mosaic}
\usage{
display_pieces(image_list)
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_bricks]{collect_bricks()}}}
}
\value{
\code{ggplot} plot object showing number of required bricks faceted by
brick color.  Each facet contains all of the unique brick types required
for that color.
}
\description{
Visualize required pieces for mosaic
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
  
display_pieces(res)
}
