% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{display_set}
\alias{display_set}
\title{Render the LEGO mosaic as a \code{ggplot}}
\usage{
display_set(image_list, title = NULL)
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_bricks]{collect_bricks()}}}

\item{title}{Optional title for plot}
}
\value{
Plot object
}
\description{
Render the LEGO mosaic as a \code{ggplot}
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
  
display_set(res)
}
