% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{generate_instructions}
\alias{generate_instructions}
\title{Produce instructions for building LEGO mosaic}
\usage{
generate_instructions(image_list, num_steps = 6)
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_bricks]{collect_bricks()}}}

\item{num_steps}{Number of steps for the instructions. Default is 6.}
}
\value{
\code{ggplot2} plot object with each facet being a particular step
}
\description{
Produce instructions for building LEGO mosaic
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
  
generate_instructions(res)
}
