% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{generate_steps}
\alias{generate_steps}
\title{Produce tidy data set for building instructions}
\usage{
generate_steps(image_list, num_steps = 6)
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_bricks]{collect_bricks()}}}

\item{num_steps}{Number of steps for the instructions. Default is 6.}
}
\value{
list with new slot containing \code{tibble} of the instructions
}
\description{
Produce tidy data set for building instructions
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
  
res_steps <- generate_steps(res)

}
