% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lego_colors}
\alias{lego_colors}
\title{LEGO colors and their R, G, and B values}
\format{A data frame (tibble) with the following columns:
\itemize{
\item \code{LEGONo}: Unique LEGO brick ID
\item \code{Color}: Color name
\item \code{R_lego}: Red color intensity proportion
\item \code{G_lego}: Green color intensity propotion
\item \code{B_lego}: Blue color intensity proportion
\item \code{c_Palette2016}: Boolean for whether color is part of 2016 palette
\item \code{c_Transparent}: Is the brick transparent?
\item \code{c_Glow}: Is the brick considered a glow brick?
\item \code{c_Metallic}: Is the brick considerend metallic?
\item \code{t_BW}: Black or white color?
\item \code{t_Classic}: Classic color?
\item \code{t_Friends}: Part of friends collection?
\item \code{w_weight}: ???
\item \code{w_Classic}: ???
}}
\usage{
lego_colors
}
\description{
A dataset containing 39 LEGO brick colors and their associated red (R), green (G),
and blue (B) color values. These colors (and more) can be found in LEGO's official
\href{http://www.bartneck.de/wp-content/uploads/2016/09/2016-LEGO-color-palette.pdf}{moulding colour palette}
(2016 version). Note that this version of the data set excludes tranparent,
glow, and metallic bricks.
}
\examples{
lego_colors
}
\keyword{datasets}
