% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{legoize}
\alias{legoize}
\title{Create LEGO-ized version of image}
\usage{
legoize(image_list, theme = "default", contrast = 1)
}
\arguments{
\item{image_list}{list object containing the \code{Img_scaled} data frame
as produced by the \code{\link[=scale_image]{scale_image()}} function.}

\item{theme}{string indicating whether to use the \code{default} color
theme or \code{bw} (grayscale) theme.}

\item{contrast}{Adjustment for color allocation when using grayscale theme.
Must be a positive number.  Default is 1, meaning no adjustment.}
}
\value{
list with modified \code{Img_lego} object (the original image lego-ized)
}
\description{
Create LEGO-ized version of image
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize()
}
