% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{list_to_li}
\alias{list_to_li}
\alias{list_to_p}
\alias{named_to_li}
\title{Turn an R list into an HTML list}
\usage{
list_to_li(list, class = NULL)

list_to_p(list, class = NULL)

named_to_li(list, class = NULL)
}
\arguments{
\item{list}{An R list}
}
\value{
an HTML list
}
\description{
Turn an R list into an HTML list
}
\examples{
list_to_li(c("a","b"))
}
