% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_choose_step.R
\name{mod_choose_stepui}
\alias{mod_choose_stepui}
\alias{mod_choose_step}
\title{mod_choose_stepui and mod_choose_step}
\usage{
mod_choose_stepui(id)

mod_choose_step(input, output, session, steps_obj, steps_inst_obj)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
