% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_define_steps.R
\name{mod_define_stepsui}
\alias{mod_define_stepsui}
\alias{mod_define_steps}
\title{mod_define_stepsui and mod_define_steps}
\usage{
mod_define_stepsui(id)

mod_define_steps(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
