% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_image.R
\name{mod_display_imageui}
\alias{mod_display_imageui}
\alias{mod_display_image}
\title{mod_display_imageui and mod_display_image}
\usage{
mod_display_imageui(id, height_window = 600)

mod_display_image(input, output, session, img_processed, width = 600,
  height = 600)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
