% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_instructions.R
\name{mod_display_instructionsui}
\alias{mod_display_instructionsui}
\alias{mod_display_instructions}
\title{mod_display_instructionsui and mod_display_instructions}
\usage{
mod_display_instructionsui(id, height_window = 500)

mod_display_instructions(input, output, session, steps_obj, step_choice)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
