% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_lego_2d.R
\name{mod_display_lego_2dui}
\alias{mod_display_lego_2dui}
\alias{mod_display_lego_2d}
\title{mod_display_lego_2dui and mod_display_lego_2d}
\usage{
mod_display_lego_2dui(id, height_window = 500)

mod_display_lego_2d(input, output, session, image_lego_obj)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
