% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_pieces.R
\name{mod_display_piecesui}
\alias{mod_display_piecesui}
\alias{mod_display_pieces}
\title{mod_display_piecesui and mod_display_pieces}
\usage{
mod_display_piecesui(id, height_window = 500)

mod_display_pieces(input, output, session, steps_obj, step_choice)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
