% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_instructions.R
\name{mod_instructionsui}
\alias{mod_instructionsui}
\alias{mod_instructions}
\title{mod_instructionsui and mod_instructions}
\usage{
mod_instructionsui(id)

mod_instructions(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
