% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mosaic_metrics.R
\name{mod_mosaic_metricsui}
\alias{mod_mosaic_metricsui}
\alias{mod_mosaic_metrics}
\title{mod_mosaic_metricsui and mod_mosaic_metrics}
\usage{
mod_mosaic_metricsui(id)

mod_mosaic_metrics(input, output, session, scale_obj)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
