% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scale_image.R
\name{mod_scale_imageui}
\alias{mod_scale_imageui}
\alias{mod_scale_image}
\title{mod_scale_imageui and mod_scale_image}
\usage{
mod_scale_imageui(id)

mod_scale_image(input, output, session, img_processed)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
