% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_pieces.R
\name{mod_table_piecesui}
\alias{mod_table_piecesui}
\alias{mod_table_pieces}
\title{mod_table_piecesui and mod_table_pieces}
\usage{
mod_table_piecesui(id)

mod_table_pieces(input, output, session, steps_obj, step_choice,
  width_table = "100\%")
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
