% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{plot_instructions}
\alias{plot_instructions}
\title{Display plot with instructions at all or specific step}
\usage{
plot_instructions(image_list, step_id = NULL)
}
\arguments{
\item{step_id}{Optional index of step to print. If \code{NULL}
(default), all steps are plotted.}

\item{steps_df}{\code{tibble} of the bricks required at each step}
}
\description{
Display plot with instructions at all or specific step
}
