% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_data}
\alias{price_data}
\title{LEGO brick price data}
\format{A data frame (tibble) with the following columns:
\itemize{
\item \code{type}: Type of brick
\item \code{size}: Dimensions in format \code{a x b} measured in studs
\item \code{shape}: Either square (\code{a x a}) or rectangle (\code{a x b})
\item \code{price}: Price for one unit of the specific brick (in U.S. dollars)
}}
\usage{
price_data
}
\description{
A dataset with prices of plate bricks for the common sizes used in building the
LEGO mosaics. These prices were derived from the \href{https://shop.lego.com}{official LEGO shop store}
as of early 2019.
}
\examples{
price_data
}
\keyword{datasets}
