% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{rep_br}
\alias{rep_br}
\title{Repeat tags$br}
\usage{
rep_br(times = 1)
}
\arguments{
\item{times}{the number of br to return}
}
\value{
the number of br specified in times
}
\description{
Repeat tags$br
}
\examples{
rep_br(5)

}
