% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{scale_image}
\alias{scale_image}
\title{Create a scaled version of jpeg image}
\usage{
scale_image(image, img_size, brightness = 1, warhol = 1:3)
}
\arguments{
\item{image}{raster array of jpeg image file as produced by \code{jpeg::readJPEG}.}

\item{img_size}{dimensions of the mosaic base plate. For a square base, supply
a single integer. For a rectangle base, supply a vector with two integers in
the form of \code{c(width, height}).}

\item{brightness}{Brightness adjustment threshold. For a darker mosaic, enter
a number between 0 and 1. For a brighter mosaic, enter a number greater than 1.
Default is \code{1}, meaning no adjustment.}

\item{warhol}{??? Default is \code{1:3}}
}
\value{
list with object called \code{Img_scaled} representing the scaled image.
}
\description{
Create a scaled version of jpeg image
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48)
}
