% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{step_pieces}
\alias{step_pieces}
\title{Generate pieces data frame categorized by each step}
\usage{
step_pieces(steps_df)
}
\arguments{
\item{steps_df}{data frame produced by \code{\link[=generate_steps]{generate_steps()}}}
}
\value{
data frame with all pieces required for each step
}
\description{
Generate pieces data frame categorized by each step
}
