% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{table_pieces}
\alias{table_pieces}
\title{Create printer-friendly version of required bricks}
\usage{
table_pieces(image_list)
}
\arguments{
\item{image_list}{List of objects produced by \code{\link[=collect_bricks]{collect_bricks()}}}
}
\value{
data frame (tibble) with columns for each brick dimension and rows
for each unique brick color in mosaic.
}
\description{
Create printer-friendly version of required bricks
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
  
table_pieces(res)
}
