% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{tagRemoveAttributes}
\alias{tagRemoveAttributes}
\title{Remove a tag attribute}
\usage{
tagRemoveAttributes(tag, ...)
}
\arguments{
\item{tag}{the tag}

\item{...}{the attributes to remove}
}
\value{
a new tag
}
\description{
Remove a tag attribute
}
\examples{
\dontrun{
a <- tags$p(src = "plop", "pouet")
tagRemoveAttributes(a, "src")
}
}
