% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{with_red_star}
\alias{with_red_star}
\title{Add a red star at the end of the text}
\usage{
with_red_star(text)
}
\arguments{
\item{text}{the HTLM text to put before the red star}
}
\value{
an html element
}
\description{
Adds a red star at the end of the text
(for example for indicating mandatory fields).
}
\examples{
with_red_star("Enter your name here")

}
