% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{collect_bricks}
\alias{collect_bricks}
\title{Collect required bricks for mosaic}
\usage{
collect_bricks(image_list, mosaic_type = "flat", num_steps = NULL)
}
\arguments{
\item{image_list}{list object containing the \code{Img_scaled} data frame
as produced by the \code{\link[=legoize]{legoize()}} function.}

\item{mosaic_type}{String for mosaic type to produce. A \code{flat} mosaic
places a single layer of LEGO plates on a base plate with study-side up.
A \code{stacked} mosaic staggers bricks and places them horizontally.}

\item{Number}{of steps}
}
\value{
list with following components:
\itemize{
\item \code{Img_bricks}: Updated LEGO image
\item \code{ID_bricks}: Data frame with brick IDs
\item \code{mosaic_type}: Type of mosaic specfied by user
\item \code{pieces}: Data frame with number of bricks needed`
}
}
\description{
Collect required bricks for mosaic
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
}
