% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{convert_to_lego_colors}
\alias{convert_to_lego_colors}
\title{Convert raw R,G,B values to LEGO version}
\usage{
convert_to_lego_colors(R, G, B, lego_colors)
}
\arguments{
\item{R}{Red color intensity value}

\item{G}{Green color intensity value}

\item{B}{Blue color intensity value}
}
\value{
data frame
}
\description{
Replaces a color defined by RGB values to a simular color in the
\code{lego_colors} data set using the shortest Euclidean distance.
}
\examples{
data(lego_colors, package = "shinylego")
new_colors <- convert_to_lego_colors(R = 30, G = 74, B = 38, lego_colors)
}
