% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_graphic.R
\name{mod_upload_graphicui}
\alias{mod_upload_graphicui}
\alias{mod_upload_graphic}
\title{mod_upload_graphicui and mod_upload_graphic}
\usage{
mod_upload_graphicui(id, label_text = NULL)

mod_upload_graphic(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{label_text}{Optional string for \code{\link[shiny:fileInput]{shiny::fileInput()}} widget}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
