% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{plot_instructions}
\alias{plot_instructions}
\title{Display plot with instructions at all or specific step}
\usage{
plot_instructions(steps_df, mosaic_type = "flat", step_id = NULL)
}
\arguments{
\item{steps_df}{data frame produced by \code{\link[=generate_steps]{generate_steps()}}}

\item{step_id}{Optional index of step to print. If \code{NULL}
(default), all steps are plotted.}
}
\description{
Display plot with instructions at all or specific step
}
