% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_functions.R
\name{table_pieces}
\alias{table_pieces}
\title{Create printer-friendly version of required bricks}
\usage{
table_pieces(pieces_df, step_id = NULL)
}
\arguments{
\item{pieces_df}{Data frame with mosaic pieces information. If summarizing
the entire mosiac, use the \code{pieces} data frame produced by \code{\link[=collect_bricks]{collect_bricks()}}.
If using a specific step, use the data frame produced by \code{\link[=step_pieces]{step_pieces()}}}

\item{step_id}{Optional step ID to subset}
}
\value{
data frame (tibble) with columns for each brick dimension and rows
for each unique brick color in mosaic.
}
\description{
Create printer-friendly version of required bricks
}
\examples{
img_file <- system.file("images", "kde_konqi_mascot.jpg", package = "shinylego")
image <- jpeg::readJPEG(img_file)
res <- scale_image(image, img_size = 48) \%>\%
  legoize() \%>\%
  collect_bricks()
  
table_pieces(res)
}
